+++ packages/communication/src/wallet/index.ts
@@ -1,4 +1,4 @@
 export type { ChainConfig } from '@cere-wallet/wallet-engine';
-export type { UserInfo, LoginData, AppContext } from './channels';
+export type { UserInfo, LoginData, LoginOptions, AppContext } from './channels';
 
 export * from './createWalletConnection';