+++ packages/embed-wallet/src/EmbedWallet.ts
@@ -1,5 +1,4 @@
 import EventEmitter from 'events';
-import { Substream } from '@toruslabs/openlogin-jrpc';
 import Torus, { TORUS_BUILD_ENV_TYPE, preloadIframe } from '@cere/torus-embed';
 import BN from 'bn.js';
 
@@ -61,6 +60,12 @@ export class EmbedWallet {
   private defaultContext: Context;
   private proxyProvider: ProxyProvider;
   private connectOptions: WalletConnectOptions = {};
+  private onAfterInit?: (error?: any) => void;
+
+  /**
+   * @description Promise that resolves when the wallet instance is initialized and ready
+   */
+  readonly isReady: Promise<EmbedWallet>;
 
   constructor({ env, clientVersion = WALLET_CLIENT_VERSION, ...options }: WalletOptions = {}) {
     if (env) {
@@ -72,6 +77,10 @@ export class EmbedWallet {
     this.proxyProvider = new ProxyProvider();
     this.defaultContext = createContext();
     this.options = { ...options, clientVersion, env: env || 'prod' };
+
+    this.isReady = new Promise((resolve, reject) => {
+      this.onAfterInit = (error) => (error ? reject(error) : resolve(this));
+    });
 
     this.provider.on('message', this.handleEvenets);
   }
@@ -114,7 +123,13 @@ export class EmbedWallet {
   }
 
   private get contextStream() {
-    return this.torus.communicationMux.getStream('app_context') as Substream;
+    const stream = this.torus.communicationMux.getStream('app_context');
+
+    if (typeof stream === 'symbol') {
+      throw new Error('Symbol streams are not supported');
+    }
+
+    return stream;
   }
 
   subscribe(eventName: WalletEvent, listener: (...args: any[]) => void) {
@@ -140,21 +155,30 @@ export class EmbedWallet {
 
     const { sessionId } = getAuthRedirectResult();
 
-    await this.torus.init({
-      network,
-      sessionId,
-      popupMode,
-      context: this.defaultContext,
-      buildEnv: buildEnvMap[env],
-      enableLogging: env !== 'prod',
-      integrity: {
-        check: false,
-        version: clientVersion,
-      },
-    });
+    try {
+      await this.torus.init({
+        network,
+        sessionId,
+        popupMode,
+        context: this.defaultContext,
+        buildEnv: buildEnvMap[env],
+        enableLogging: env !== 'prod',
+        integrity: {
+          check: false,
+          version: clientVersion,
+        },
+      });
 
-    this.proxyProvider.setTarget(this.torus.provider);
-    this.setStatus(this.torus.isLoggedIn ? 'connected' : 'ready');
+      this.proxyProvider.setTarget(this.torus.provider);
+      this.setStatus(this.torus.isLoggedIn ? 'connected' : 'ready');
+
+      this.onAfterInit?.();
+    } catch (error) {
+      this.setStatus('errored');
+      this.onAfterInit?.(error);
+
+      throw error;
+    }
   }
 
   async connect(overrideOptions: WalletConnectOptions = {}) {