+++ src/constants.ts
@@ -25,3 +25,9 @@ export const FEATURE_FLAGS = {
  * TODO: Turn it back when fixed for mobiles
  */
 export const SUPPORTED_SOCIAL_LOGINS: ('google' | 'facebook')[] = [];
+
+export const COIN_TICKER_ALIAS: Record<string, string> = {
+  CERE_ERC20: 'CERE',
+};
+
+export const RPC_POLLING_INTERVAL = 10000;