+++ src/routes/index.tsx
@@ -1,8 +1,2 @@
-import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-
-import routes from './routes';
-
-const router = createBrowserRouter(routes);
-
 export * from './RouteElement';
-export const Router = () => <RouterProvider router={router} />;
+export { AppRouter as Router } from './AppRouter';