+++ src/stores/WalletStore/WalletStore.ts
@@ -16,7 +16,7 @@ import { CollectiblesStore } from '../CollectiblesStore';
 import { OpenLoginStore } from '../OpenLoginStore';
 import { ApprovalStore } from '../ApprovalStore';
 import { PopupManagerStore } from '../PopupManagerStore';
-import { CERE_NETWORK_RPC } from '~/constants';
+import { CERE_NETWORK_RPC, RPC_POLLING_INTERVAL } from '~/constants';
 import { ApplicationsStore } from '../ApplicationsStore';
 
 export class WalletStore implements Wallet {
@@ -114,28 +114,31 @@ export class WalletStore implements Wallet {
     }
 
     const engine = createWalletEngine({
+      pollingInterval: RPC_POLLING_INTERVAL,
       chainConfig: this.network!,
       polkadotRpc: CERE_NETWORK_RPC,
-      getAccounts: () => this.accountStore.accounts,
+      getAccounts: (pairs) => this.accountStore.mapAccounts(pairs),
+      onUpdateAccounts: (accounts) => this.accountStore.updateAccounts(accounts),
       getPrivateKey: () => this.accountStore.privateKey,
       onPersonalSign: (request) => this.approvalStore.approvePersonalSign(request),
       onSendTransaction: (request) => this.approvalStore.approveSendTransaction(request, { showDetails: true }),
       onTransfer: (request) => this.approvalStore.approveTransfer(request),
     });
 
+    await engine.updateAccounts();
+
     runInAction(() => {
       this.currentEngine = engine;
       this.currentProvider = new providers.Web3Provider(engine.provider);
+
+      this.currentProvider.pollingInterval = RPC_POLLING_INTERVAL;
 
       this.initialized = true;
     });
 
     reaction(
-      () => this.accountStore.accounts,
-      (accounts) => engine.updateAccounts(accounts),
-      {
-        fireImmediately: true,
-      },
+      () => this.accountStore.privateKey,
+      () => engine.updateAccounts(),
     );
   }
 }