+++ packages/embed-wallet/CHANGELOG.md
@@ -3,3 +3,5 @@
 ### vNext
 
 - Add CERE ERC20 balance and transfer support
+- Add `isReady` promise property to the wallet instance
+- Fix account changed event delay after wallet connect